Added Update feature, Finished Options.
The Update feature is now working and apart of Aisis 1.2, All options are fully implemented and working.

+++ CoreTheme/Update.php
@@ -2,8 +2,14 @@
 
 class CoreTheme_Update implements AisisCore_Interfaces_Upgrade{
 	
+	protected $_xml_object;
+	
+	public function __construct(){
+		$this->_xml_object = simplexml_load_file('http://adambalan.com/aisis/aisis_update/aisis_version.xml');
+	}
+	
 	public function check_for_update(){
-		if($this->check_theme_version() != ''){
+		if($this->check_theme_version() != '' && $this->get_update_notice() == 'yes'){
 			if(version_compare($this->check_theme_version(), $this->get_current_version(), '>')){
 				return true;
 			}		
@@ -13,47 +19,49 @@ class CoreTheme_Update implements AisisCore_Interfaces_Upgrade{
 	public function upgrade(){
 		global $wp_filesystem;
 		
-		if(current_user_can('update_themes')){
-			$this->cred_check();
-
-			$aisis_temp_file_download = download_url( 'http://adambalan.com/aisis/aisis_update/Aisis.zip' );
-
-			if(is_wp_error($aisis_temp_file_download)){
-				$error = $aisis_temp_file_download->get_error_code();
-				throw new CoreTheme_Exception_UpdateException($aisis_temp_file_download->get_error_message($error));
+		if(current_user_can('update_themes')){				
+			if($this->_cred_check()){
+				request_filesystem_credentials(admin_url('admin.php?page=aisis-core-update'));
+			}else{
+				$aisis_temp_file_download = download_url( 'http://adambalan.com/aisis/aisis_update/AisisUpdate.zip' );
+	
+				if(is_wp_error($aisis_temp_file_download)){
+					throw new CoreTheme_Exceptions_UpdateException('We encounterd an issue and could not continue: ' .$aisis_do_unzip->get_error_code());
+				}
+				
+				$aisis_unzip_to = $wp_filesystem->wp_content_dir() . "/themes/Aisis";
+				$aisis_do_unzip = unzip_file($aisis_temp_file_download, $aisis_unzip_to);
+				
+				unlink($aisis_temp_file_download);
+				
+				if(is_wp_error($aisis_do_unzip)){
+					throw new CoreTheme_Exceptions_UpdateException('We encounterd an issue and could not continue: ' .$aisis_do_unzip->get_error_code());
+				}
+				
+				$http = new AisisCore_Http_Http();
+				wp_safe_redirect(admin_url('index.php?aisis_upgrade=true'));
 			}
-			
-			$aisis_unzip_to = $wp_filesystem->wp_content_dir() . "/themes/Aisis";
-			$aisis_do_unzip = unzip_file($aisis_temp_file_download, $aisis_unzip_to);
-			
-			unlink($aisis_temp_file_download);
-			
-			if(is_wp_error($aisis_do_unzip)){
-				$error = $aisis_do_unzip->get_error_code();
-				throw new CoreTheme_Exception_UpdateException($aisis_temp_file_download->get_error_message($error));
-			}
-			
-			$http = new AisisCore_Http_Http();
-			wp_safe_redirect(admin_url('admin.php?page=aisis-core-options&noheader=true&upgrade=true'));
 		}		
 	}
 	
 	protected function _cred_check(){
 		$aisis_file_system_structure = WP_Filesystem();
 		if($aisis_file_system_structure == false){
-			add_action('admin_notices', array($this, 'need_credentials'));
+			echo '<div class="alert"><strog>NOTE!</strong> We need your ftp creds before we continue!</div>';
 			return true;
-		}
-
-		return false;			 
+		}		 
+		
+		return false;
 	}	
 	
 	public function check_theme_version(){
-		$version_url = 'http://adambalan.com/aisis/version.xml';
-		$aisis_update_xml_object = simplexml_load_file($version_url);
-		$aisis_version = $aisis_update_xml_object->version[0];
-		
+		$aisis_version = $this->_xml_object->version[0];
 		return trim($aisis_version);	
+	}
+	
+	public function get_update_notice(){
+		$update = $this->_xml_object->update[0];
+		return trim($update);			
 	}
 	
 	public function get_current_version(){
